Worker cleanup, removed database connection retry
