+++ plenario_worker/clients.py
@@ -1,0 +1,17 @@
+import boto.sqs
+import boto3
+from plenario.settings import AWS_ACCESS_KEY, AWS_SECRET_KEY, AWS_REGION_NAME
+
+
+sqs_client = boto.sqs.connect_to_region(
+    AWS_REGION_NAME,
+    aws_access_key_id=AWS_ACCESS_KEY,
+    aws_secret_access_key=AWS_SECRET_KEY
+)
+
+autoscaling_client = boto3.client(
+    'autoscaling',
+    region_name=AWS_REGION_NAME,
+    aws_access_key_id=AWS_ACCESS_KEY,
+    aws_secret_access_key=AWS_SECRET_KEY
+)
